[WIP] More fixes (#7578)
- CUDA uses tad cache (and new tads) now
- INDArray getRow/getColumn now use tensorAlongDimension()
- pullRows test fix + additional validation

+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ndarray/INDArray.java
@@ -158,7 +158,7 @@ public interface INDArray extends Serializable, AutoCloseable {
      * @param dimension the dimension to getScalar the vector from
      * @return the vector along a particular dimension
      */
-    INDArray tensorAlongDimension(int index, int... dimension);
+    INDArray tensorAlongDimension(long index, int... dimension);
 
     /**
      * Get the vector along a particular dimension
